feat: remove 'type' from package.json
Laravel added default "type": "module" to their package.json in PR: https://github.com/laravel/vite-plugin/pull/189. This breaks compilation with laravel-mix and therefore we need to remove this line from package.json in the installation process. The ability to compile with Vite shouldn't be affected by this, however some specific modules (like svelte) can be affected by this, as mentioned in the Laravel's PR. We think that the potential problem with specific libraries should be resolved by developer that uses them.